refactor(server,common): remove `writing` database field
The idea behind the `writing` field was to prevent stuff from
finishing/expiring the upload while a chunk was in progress.
However, decrementing it is a database call that can fail
(and if it fails, we can't finish or expire the upload).

It's also been replaced by the `flock`ing code, which is more
reliable anyway.